Merge EnTT implementation with dev

+++ Sail/src/Sail/debug/Instrumentor.h
@@ -149,7 +149,7 @@ public:
 
 		m_result.start = std::chrono::time_point_cast<std::chrono::microseconds>(m_startTimepoint).time_since_epoch().count();
 		m_result.end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
-		m_result.threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+		m_result.threadID = static_cast<uint32_t>(std::hash<std::thread::id>{}(std::this_thread::get_id()));
 		Instrumentor::Instance().writeProfile(m_result);
 
 		m_stopped = true;
